requirements.txt
openai
python-dotenv
pydub
# webvtt-py